feat: introduce international-types
