+++ packages/next-international/src/i18n/create-define-locale.ts
@@ -1,4 +1,4 @@
-import { BaseLocale, LocaleValue } from '../types';
+import { BaseLocale, LocaleValue } from 'international-types';
 
 export function createDefineLocale<Locale extends BaseLocale>() {
   return function defineLocale(locale: { [key in keyof Locale]: LocaleValue }) {
