+++ packages/next-international/src/i18n/create-i18n-provider.tsx
@@ -1,5 +1,6 @@
 import React, { Context, ReactElement, ReactNode, useEffect, useState } from 'react';
-import { LocaleContext, Locales, BaseLocale } from '../types';
+import { LocaleContext, Locales } from '../types';
+import { BaseLocale } from 'international-types';
 import { useRouter } from 'next/router';
 import { error, warn } from '../helpers/log';
