+++ packages/next-international/src/i18n/create-use-i18n.ts
@@ -1,14 +1,6 @@
 import React, { useContext, Context } from 'react';
-import {
-  BaseLocale,
-  LocaleContext,
-  LocaleKeys,
-  LocaleValue,
-  Params,
-  ParamsObject,
-  ScopedValue,
-  Scopes,
-} from '../types';
+import { BaseLocale, LocaleKeys, LocaleValue, Params, ParamsObject, ScopedValue, Scopes } from 'international-types';
+import { LocaleContext } from '../types';
 
 export function createUsei18n<Locale extends BaseLocale>(I18nContext: Context<LocaleContext<Locale> | null>) {
   return function useI18n() {
