+++ packages/next-international/src/index.tsx
@@ -1,3 +1,2 @@
 export { createI18n } from './i18n/create-i18n';
-/* c8 ignore next */
-export { BaseLocale, LocaleValue } from './types';
+export { BaseLocale, LocaleValue } from 'international-types';
